requirements.txt
google.generativeai
seaborn
grpcio==1.59.2
urllib3==1.26.7
streamlit
pandas
pandasai==2.0.43
openai
pyyaml
matplotlib
numpy
openpyxl
scikit-learn
langchain
langchain_community
ollama
SpeechRecognition
vosk
soundfile
whisper-mic
audio_recorder_streamlit
streamlit-float
sounddevice
openai-whisper
srt
tdqm
evdev
